# ENH: linalg.solve: use structure to speed up finite check, apply `rcond` estimates