#### Reference issue
gh-12824

#### What does this implement/fix?
This has a few enhancements:
- 3cc3d1324f611ac7d18110b22be0ead8987e2034 waits until after the structure is determined to perform the finiteness check. This provides a big speedup for diagonal and tridiagonal matrices
- 361527bf302d256c57dac02fc7c698a90446e904 and b9414fe6a7a614d3a9daeb08e7bd8b29a3bac036 add a reciprocal condition number check for triangular and tridiagonal matrices respectively.
- After gh-21352 merges, I'd like to use `dlantr` to do the norm calculation for triangular matrices as part of this PR.
 
#### Additional information
Time information for solving `linalg.solve(A, b)` with 1000x1000 matrix:
```
# diagonal
# before: 995 μs ± 21.6 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# after: 726 μs ± 6.73 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

# tridiagonal
# before: 1.29 ms ± 20.9 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# after finiteness check improvement: 778 μs ± 38.8 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# after adding use of trcon: 874 μs ± 5.2 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

# lower triangular
# before: 5.49 ms ± 49.7 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# after trcon: 9.34 ms ± 162 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# using `lantr` instead of `lange` gives us about 1ms back
```