@mdhaber 2024-08-12
I was also hoping to somehow speed up solution for symmetric matrices here, but if we're going for speed, it looks like the best thing we can do is to treat them as general matrices rather than using the symmetric matrix routines. With `n = 1000`, `sytrf` takes _5~10 times_ as long as `getrf` for me , depending on OS/environment.

```python3
import numpy as np
from scipy import linalg
rng = np.random.default_rng(234859349452)

n = 1000
b = rng.random(size=n)
A = rng.random((n, n))
A += A.T

A = A.T.copy()
getrf, sytrf = linalg.get_lapack_funcs(('getrf', 'sytrf'), (A, b))
hetrf = linalg.get_lapack_funcs(('hetrf'), (A.astype(np.complex128), b))

# %timeit getrf(A)  # (on PC) 19.9 ms ± 23.1 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
# %timeit sytrf(A)  # (on PC) 213 ms ± 8.91 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

lu, piv, info = getrf(A)
ldu, ipiv, info = sytrf(A)

getrs, sytrs = linalg.get_lapack_funcs(('getrs', 'sytrs'), (A, b))

# %timeit getrs(lu, piv, b)  # 150 μs ± 3.49 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# %timeit sytrs(ldu, ipiv, b)  # 327 μs ± 6.67 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

ref, _ = getrs(lu, piv, b)
res, _ = sytrs(ldu, ipiv, b)

np.testing.assert_allclose(res, ref)

gesv, sysv = linalg.get_lapack_funcs(('gesv', 'sysv'), (A, b))

_, _, ref, _ = gesv(A, b)
_, _, res, _ = sysv(A, b)

%timeit gesv(A, b)  # (on mac) 18.5 ms ± 2.64 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
%timeit sysv(A, b)  # (on mac) 87.7 ms ± 256 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
np.testing.assert_allclose(res, ref)
```

It's possible that it's *more* stable, I suppose? It doesn't seem more accurate for these random matrices, though, so I'm guessing that if there is an advantage, it's only for special cases. Since `solve` is for general use, what do you think about removing the specialization for symmetric matrices? (And Hermitian, too, based on limited testing.)