@ilayn 2024-08-12
Are you on OpenBLAS? Because `getrf` in OpenBLAS is rewritten in C and I think `sytrf` is not yet.