@mdhaber 2024-10-18
Thanks @ilayn!

> It gets a bit tiring to type upper triangular after a while instead of utri or whatever.

Can that be a separate PR? 
I'm ambivalent... I have liked short, obvious names, but there is a movement toward spelled-out names,. I don't really want to be the one to stoke that fire.
If so, please merge at will. I'm done here.

---

The last thing on my agenda for these backsolve structure checks was `solve_banded`. Do you want to see a PR for that before 1.15.0 branches?