@mdhaber 2024-10-19
OK, done in gh-21726.

I think I will do just one more thing, which is to open an issue about the slowdown when assuming symmetry/positive definiteness. 

Other than that, I think I will not open any more linalg PRs until the others (e.g. gh-21462, gh-21446, gh-21419) are done, but I'd be happy to review PRs with other enhancements we've discussed. Thanks!